requirements.txt
pyjnius==1.3.0
